requirements.txt
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
click==8.2.1
ecdsa==0.19.1
fastapi==0.115.14
greenlet==3.2.3
h11==0.16.0
idna==3.10
mako==1.3.10
markupsafe==3.0.2
passlib==1.7.4
pillow==11.3.0
pyasn1==0.6.1
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.41
starlette==0.46.2
typing-extensions==4.14.1
typing-inspection==0.4.1
uvicorn==0.35.0
